+++ monitoring/influxdb.go
@@ -6,10 +6,9 @@ import (
 	"strings"
 	"time"
 
+	influxdb "github.com/influxdata/influxdb1-client/v2"
 	"github.com/pkg/errors"
 	"github.com/theplant/appkit/log"
-
-	influxdb "github.com/influxdata/influxdb/client/v2"
 )
 
 // InfluxMonitorConfig type for configuration of Monitor that sinks to
@@ -75,7 +74,7 @@ func NewInfluxdbMonitor(config InfluxMonitorConfig, logger log.Logger) (Monitor,
 
 	// check connectivity to InfluxDB every 5 minutes
 	go func() {
-		t := time.NewTimer(5 * time.Minute)
+		t := time.NewTicker(5 * time.Minute)
 
 		for {
 			// Ignore duration, version